add os arquivos e updates do projeto - 23/08/23

+++ lib/src/pages/cart/view/components/cart_tile.dart
@@ -64,7 +64,7 @@ class _CartTileState extends State<CartTile> {
           result: (quantity) {
             controller.changeItemQuantity(
               item: widget.cartItem,
-              quantity: quantity as int,
+              quantity: quantity,
             );
           },
           isRemovable: true,
